requirements.txt
tensorflow==2.13.0
numpy==1.24.3
flask==2.3.3
pillow==10.0.0
matplotlib==3.7.2
scikit-learn==1.3.0
opencv-python==4.8.0.76
gunicorn==20.1.0
reportlab==4.0.4 